Refactor UserRoutes MM-314
